Setting up the project for the inclusion of pretrained models
