+++ pyproject.toml
@@ -25,6 +25,7 @@ transformers = "^4.41.0"
 sacrebleu = "^2.4.2"
 sentencepiece = "^0.2.0"
 accelerate = "^0.30.1"
+spacy = "^3.7.4"
 
 
 [build-system]
